Core/src/main/java/com/songoda/core/utils/LocationUtils.java
package com.songoda.core.utils;

import org.bukkit.Location;

public class LocationUtils {
    /**
     * Compares the block coordinates of two locations <strong>ignoring the world</strong>
     */
    public static boolean isLocationMatching(Location location1, Location location2) {
        return location1.getBlockX() == location2.getBlockX() &&
                location1.getBlockY() == location2.getBlockY() &&
                location1.getBlockZ() == location2.getBlockZ();
    }

    public static boolean isInArea(Location l, Location pos1, Location pos2) {
        double x1 = Math.min(pos1.getX(), pos2.getX());
        double y1 = Math.min(pos1.getY(), pos2.getY());
        double z1 = Math.min(pos1.getZ(), pos2.getZ());

        double x2 = Math.max(pos1.getX(), pos2.getX());
        double y2 = Math.max(pos1.getY(), pos2.getY());
        double z2 = Math.max(pos1.getZ(), pos2.getZ());

        return l.getX() >= x1 && l.getX() <= x2 &&
                l.getY() >= y1 && l.getY() <= y2 &&
                l.getZ() >= z1 && l.getZ() <= z2;
    }
}
